Hashmap find unique Two-Sum Pairs
Design an algorithm for the two-sum problem using HashMap. Use the built in Java API’s for the HashMap class. You have to count the number of all the unique pairs possible by picking two elements from the array which sum up to the given target sum. The input will consist of three lines:
The size of the array
Elements of the array (space-separated)
The target sum
Use a HashMap whose key is the element of the array and value determines if the occurrence of that element is single or multiple. You have to output the number of distinct pairs possible. Print 0 if no such pairs are possible. Remember the ordering in the pair doesn’t matter, so (1,3) and (3,1) will be considered the same pair and not two distinct pairs. Complete the main method and the getPairsCount method accordingly.